requirements.txt
requests>=2.31.0,<2.32.0
python-dotenv>=1.0.0,<1.1.0